Move to HTTPFormatSequence.
